[refactor]: rename BlogDetail component to BlogCard and update BlogList with dynamic blog data
