+++ src/assets/css/main.css
@@ -72,4 +72,34 @@
     color: var(--heading-color);
     font-weight: 500;
   }
+
+  ::selection {
+    @apply bg-emerald-600 text-white;
+  }
+
+  /*  Custom Scrollbar */
+  /* Toàn site dùng scrollbar mảnh và cùng tone màu */
+  * {
+    scrollbar-width: thin; /* Firefox */
+    scrollbar-color: theme('colors.indigo.400') transparent;
+  }
+
+  /* Chrome, Edge, Safari */
+  *::-webkit-scrollbar {
+    width: 10px;
+    height: 10px;
+  }
+
+  *::-webkit-scrollbar-track {
+    background: transparent;
+  }
+
+  *::-webkit-scrollbar-thumb {
+    background-color: theme('colors.indigo.400');
+    border-radius: 9999px;
+  }
+
+  *::-webkit-scrollbar-thumb:hover {
+    background-color: theme('colors.indigo.500');
+  }
 }
