+++ src/components/blogs/BlogCard.vue
@@ -1,0 +1,5 @@
+<script setup></script>
+
+<template></template>
+
+<style scoped></style>
